% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsmartr_unstack.R
\name{dsmartr_class_maps}
\alias{dsmartr_class_maps}
\title{Produce soil class probability surfaces}
\usage{
dsmartr_class_maps(tallied_probs = NULL, soil_class = NULL, lookup = NULL,
  cpus = 1)
}
\arguments{
\item{tallied_probs}{RasterBrick; \code{tallied_probabilities} output by
\code{\link{dsmartr_collate}}.}

\item{soil_class}{String; Soil class(es) of interest, if only particular maps are desired.}

\item{lookup}{Data Frame; contains raster values and corresponding soil class labels. See
\code{\link{dsmartr_collate}}.}

\item{cpus}{Integer; number of processors to use in parallel.}
}
\value{
\code{class_maps}: List of RasterLayers; probability surfaces for each soil class.

All outputs are written to disk as GeoTIFFs before being assigned to the global environment.
}
\description{
Produce soil class probability surfaces
}
\examples{
\dontrun{
# run dsmartr_collate() with the example data and then

# all classes:
dsmartr_class_maps(tallied_probs = tallied_probabilities, lookup = LUT)

# just map two classes of interest:
dsmartr_class_maps(tallied_probs = tallied_probabilities,
soil_class = c('BL', 'CO'), lookup = LUT)}
}
