% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsmartr_evaluate.R
\name{dsmartr_eval_npred}
\alias{dsmartr_eval_npred}
\title{Calculate soil classes per pixel}
\usage{
dsmartr_eval_npred(tallied_preds = NULL, cpus = 1)
}
\arguments{
\item{tallied_preds}{RasterBrick; \code{tallied_predictions} output by
\code{\link{dsmartr_collate}}.}

\item{cpus}{Integer; number of processors to use in parallel.}
}
\value{
\code{n_classes_predicted}: RasterLayer depicting the number of distinct soils predicted
  per pixel. Written to disk as GeoTIFF.
}
\description{
Calculates and maps the number of distinct soil classes predicted by dsmartr across \code{n}
iterations. Requires output from \code{\link{dsmartr_collate}}.
}
\note{
Fewer classes predicted on a pixel generally indicates higher internal model confidence at
  that location.
}
\examples{
\dontrun{
# run dsmartr_iterate() and dsmartr_collate() with the example data then:
npred <- dsmartr_eval_npred(tallied_preds = collated[['tallied_predictions']],
cpus = max(1, (parallel::detectCores() - 1)))}
}
