% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsmartr_evaluate.R
\name{dsmartr_eval_ties}
\alias{dsmartr_eval_ties}
\title{Detect ties for most-probable soil}
\usage{
dsmartr_eval_ties(tallied_preds = NULL, cpus = 1)
}
\arguments{
\item{tallied_preds}{RasterBrick; \code{tallied_predictions} output by
\code{\link{dsmartr_collate}}.}

\item{cpus}{Integer; number of processors to use in parallel.}
}
\value{
\code{n_classes_predicted}: RasterLayer depicting the number of soils tied for
  most-probable per pixel. Written to disk as GeoTIFF.
}
\description{
Calculates and maps out locations where the most-probable soil is a tie between two or more
classes. Requires output from \code{\link{dsmartr_collate}}.
}
\examples{
\dontrun{
# run dsmartr_iterate() and dsmartr_collate() with the example data then:
tiemap <- dsmartr_eval_ties(tallied_preds = collated[['tallied_predictions']],
cpus = max(1, (parallel::detectCores() - 1)))}
}
