% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsmartr_iterate.R
\name{dsmartr_iterate}
\alias{dsmartr_iterate}
\title{dsmartr model iterator}
\usage{
dsmartr_iterate(prepped_map = NULL, covariates = NULL,
  prepped_points = NULL, id_field = NULL, n_iterations = NULL,
  t_factor = 10000, c5_ctrl = NULL, cpus = 1,
  write_files = c("rds_only", "all"), write_samples = FALSE,
  resume_from = NULL)
}
\arguments{
\item{prepped_map}{Data Frame; Returned by \code{\link{dsmartr_prep_polygons}}.}

\item{covariates}{RasterStack or RasterBrick; environmental covariate datasets.}

\item{prepped_points}{Data Frame; Returned by \code{\link{dsmartr_prep_points}}; optional.}

\item{id_field}{String; name of unique polygon identifier field.}

\item{n_iterations}{Integer; desired number of model runs.}

\item{t_factor}{Integer; dirichlet distribution modifier. Where map units have >1 soil class
component, larger values of \code{t_factor} dampen variation away from the input proportions.}

\item{c5_ctrl}{List; output of \code{\link[C50]{C5.0Control}}; optional.}

\item{cpus}{Integer; number of processors to use in parallel.}

\item{write_files}{String; choose 'all' to write outputs in both rds and 'normal' files (GeoTIFF,
ESRI shapefile, csv, txt as appropriate).}

\item{write_samples}{Logical; whether to retain the covariate samples taken during each
iteration. If \code{TRUE}, each set of samples is written to disk as rds and/or ESRI Shapefile.}

\item{resume_from}{Integer; If this function is interrupted, it can be resumed from a specified
iteration number (prevents overwriting existing model iterations in output directory).}
}
\value{
RasterStack; \code{n_iterations} layers. As a side effect, for each model run, outputs
written to disk include the predicted soils map, the C5.0 model summary, and optionally, the
covariate sample points in spatial data format.
}
\description{
Disaggregates an input soil map a given number of times.
}
\note{
This function writes a large number of files to disk.
}
\examples{
\dontrun{
## Polygons only:
# run dsmartr_prep_polygons() per the example code for that function, then
dsmartr_iterate(prepped_map = prepped_polygons, covariates = heronvale_covariates,
 id_field = 'POLY_NO', n_iterations = 20,
 cpus = max(1, (parallel::detectCores() - 1)), write_files = 'all', write_samples = TRUE)

## Oh no, there was a blackout halfway through my model run:
dsmartr_iterate(prepped_map = prepped_polygons, covariates = heronvale_covariates,
id_field = 'POLY_NO', n_iterations = 6, cpus = max(1, (parallel::detectCores() - 1)),
write_files = 'all', write_samples = TRUE, resume_from = 14)}
}
