% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsmartr_unstack.R
\name{dsmartr_most_likely}
\alias{dsmartr_most_likely}
\title{Extract most-likely-soil maps}
\usage{
dsmartr_most_likely(dsmart_preds = NULL, dsmart_probs = NULL, n_maps = 2)
}
\arguments{
\item{dsmart_preds}{RasterBrick; output \code{dsmart_predictions} from
\code{\link{dsmartr_collate}}.}

\item{dsmart_probs}{RasterBrick; output \code{dsmart_probabilities} from \code{\link{dsmartr_collate}};
optional.}

\item{n_maps}{integer; the number of most-probable maps desired.}
}
\value{
A list containing:
\itemize{
  \item{\code{most_likely_maps}: A list of \code{n_maps} most-likely-soils maps, from most to least
  probable.}
  \item{\code{most_likely_ps}: Optional; A list of \code{n_maps} probability surfaces
  associated with \code{most_likely_maps}.}}
All outputs are written to disk as GeoTIFFs.
}
\description{
This function takes the outputs of \code{\link{dsmartr_collate}} and extracts the top \code{n}
most-likely soil maps. Optionally, their respective probability surfaces are also extracted.
}
\examples{
\dontrun{
# run dsmartr_collate() with the example data and then:
most_likely <- dsmartr_most_likely(dsmart_preds = collated[['dsmartr_predictions']],
dsmart_probs = collated[['dsmartr_probabilities']], n_maps = 2)}
}
