% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsmartr_unstack.R
\name{dsmartr_most_probable}
\alias{dsmartr_most_probable}
\title{Extract most-probable soil maps}
\usage{
dsmartr_most_probable(dsmart_preds = NULL, dsmart_probs = NULL,
  n_maps = 2)
}
\arguments{
\item{dsmart_preds}{RasterBrick; output \code{dsmart_predictions} from
\code{\link{dsmartr_collate}}.}

\item{dsmart_probs}{RasterBrick; output \code{dsmart_probabilities} from \code{\link{dsmartr_collate}};
optional.}

\item{n_maps}{integer; the number of most-probable maps desired.}
}
\value{
Two lists of RasterLayer objects:
\itemize{
  \item{\code{most_prob_maps}: A list of \code{n_maps} most-probable maps, from most to least
  probable.}
  \item{\code{most_prob_prob_maps}: Optional; A list of \code{n_maps} probability surfaces
  associated with \code{most_prob_maps}.}}
 All outputs are written to disk as GeoTIFFs before being assigned to the global environment.
}
\description{
This function takes the outputs of \code{\link{dsmartr_collate}} and extracts the top \code{n}
most-probable soil maps. Optionally, their respective probability surfaces are also extracted.
}
\examples{
\dontrun{
# run dsmartr_collate() with the example data and then:
dsmartr_most_probable(dsmart_preds = dsmart_predictions, dsmart_probs = dsmart_probabilities,
n_maps = 2)}
}
