% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsmartr_prepare.R
\name{dsmartr_prep_points}
\alias{dsmartr_prep_points}
\title{Prepare dsmartr points}
\usage{
dsmartr_prep_points(known_points = NULL, soil_id = NULL, x_coords = NULL,
  y_coords = NULL, covariates = NULL)
}
\arguments{
\item{known_points}{sfc_POINT, SpatialPointsDataFrame or Data Frame object; represents locations
where the soil class has been directly observed. Supply with spatial location (spatial data or
x and y attributes), unique numeric ID, and soil class.}

\item{soil_id}{String; name of column in \code{known_points} holding soil class data.}

\item{x_coords}{String; name of column in \code{known_points} holding x-coordinate data.
Only needed for non-spatial inputs.}

\item{y_coords}{String; name of column holding y-coordinate data. Only needed for non-spatial
inputs.}

\item{covariates}{RasterStack or RasterBrick; environmental covariate data.}
}
\value{
A data frame with two attributes: soil class code and corresponding raster cell index
number.
}
\description{
Prepares dsmartr point inputs for use in \code{\link{dsmartr_iterate}}.
}
\note{
The output of this function is an optional input for \code{\link{dsmartr_iterate}}. If
\code{known_points} is supplied as a data frame, the x and y coordinates must match the covariate
crs.
}
\examples{
\dontrun{
load('heronvale_known_sites')
load('heronvale_soilmap')
load('heronvale_covariates')

# data frame input:
dsmartr_prep_points(src_map = heronvale_soilmap, known_points = heronvale_known_sites,
soil_id = 'CLASS', x_coords = 'x', y_coords = 'y', covariates = heronvale_covariates) }
}
