% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsmartr_prepare.R
\name{dsmartr_prep_polygons}
\alias{dsmartr_prep_polygons}
\title{Prepare dsmartr polygons}
\usage{
dsmartr_prep_polygons(src_map = NULL, covariates = NULL, id_field = NULL,
  sample_method = c("flat", "area_p"), flat_rate = NULL, area_rate = NULL,
  floor = NULL)
}
\arguments{
\item{src_map}{An sfc_POLYGON, sfc_MULTIPOLYGON, or SpatialPolygonsDataFrame object representing
the soil map to be disaggregated. Format requirements:
\itemize{
\item{one row per polygon (data is wide-formatted)}
\item{A numeric unique ID field for polygons}
\item{1-n character columns for soil classes named CLASS_1 to CLASS_n}
\item{1-n numeric columns for soil class percentages named PERC_1 to PERC_n. PERC_1 must relate
to CLASS_1, etc.}
}
Other columns may exist in the object; they will be ignored.}

\item{covariates}{RasterStack or RasterBrick; environmental covariate data.}

\item{id_field}{String; name of unique identifier field in \code{src_map}.}

\item{sample_method}{String; choice of flat rate per polygon or area-proportional rate.}

\item{flat_rate}{Integer; Number of samples per polygon; use with \code{sample_method = 'flat'}.}

\item{area_rate}{Integer; desired number of samples per square kilometre; use with
\code{sample_method = 'area_p'}.}

\item{floor}{Integer; desired minimum number of samples per polygon. Optional; use with
\code{sample_method = 'area_p'}. Defaults to the number of soil classes on a polygon.}
}
\value{
A data frame holding polygon input attributes and four new attribute columns:
\itemize{
  \item{\code{area_sqkm}: Polygon area in square kilometers, by \code{\link[sf]{st_area}}.}
  \item{\code{n_classes}: The number of soil classes within the map unit.}
  \item{\code{n_samples}: The number of environmental covariate point samples that will be taken
  on each model iteration.}
  \item{\code{intersecting_cells}: Raster cell index numbers for any cell whose center falls
  within the polygon boundary.}
  }
Outputs are also written to disk.
}
\description{
Prepares dsmartr polygon inputs for use in \code{\link{dsmartr_iterate}}.
}
\note{
\itemize{
\item{The output of this function is a required input for \code{\link{dsmartr_iterate}}}.
\item{Covariate data must be in a projected CRS with defined units (not lat/long). Vector inputs
will be transformed to match the covariate CRS.}
\item{The \code{intersecting_cells} attribute field is a list-column, so the returned  object
cannot be written to e.g. csv format.}
\item{This function runs faster with a RasterBrick than a Stack.}}
}
\examples{
\dontrun{
load("heronvale_soilmap")
load("heronvale_covariates")

# flat rate
dsmartr_prep_polygons(src_map = heronvale_soilmap, covariates = heronvale_covariates,
id_field = 'POLY_NO', sample_method = 'flat', flat_rate = 20)

# area_proportional rate with floor
dsmartr_prep_polygons(src_map = heronvale_soilmap, covariates = heronvale_covariates,
id_field = 'POLY_NO', sample_method = 'area_p', area_rate = 20, floor = 10)}
}
