% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsmartr_iterate.R
\name{get_classes}
\alias{get_classes}
\title{Get all soil classes}
\usage{
get_classes(soil_map = NULL, soil_points = NULL, col_stub = NULL)
}
\arguments{
\item{col_stub}{String; stub of attribute column names holding soil class data}

\item{soilmap}{Data Frame; Returned by \code{\link{prep_polygons}}.}

\item{soilpoints}{Data Frame; Returned by \code{\link{prep_points}}; optional.}
}
\value{
A factor holding the set of soil classes occurring in both datasets, sorted
  alphabetically. Internal to \code{\link{iterate}}.
}
\description{
Returns the set of unique soil classes that occur in an input soil map, and, optionally,
a set of known points.
}
\note{
It is up to the user to make sure the classification schema in both datasets matches/is
  compatible.
}
\examples{
\dontrun{
# run prep_polygons() and prep_points() examples, then:
class_levels <- get_classes(soil_map = pr_ap, soil_points = pr_points, col_stub = 'CLASS')}
}
\keyword{internal}
