% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{heronvale_known_sites}
\alias{heronvale_known_sites}
\title{Data: Points where soil class is known}
\format{A tbl/sf_POINT object in EPSG:3577 with 36 rows and 6 columns:
\describe{
  \item{PROJECT_CODE}{Three-letter code identifying the soil survey}
  \item{SITE_ID}{Within-project unique site identifier}
  \item{LONGITUDE}{Longitude coordinate in EPSG:4283 (GDA94)}
  \item{LATITUDE}{Latitude coordinate in EPSG:4283 (GDA94)}
  \item{CLASS}{Soil class identified at this site}
  \item{geom}{list-column with class 'sfc_GEOMETRY', 'sfc'}}
Site locations are considered accurate to within +/- 50 meters.}
\source{
\itemize{
  \item{\url{https://publications.qld.gov.au/dataset/soils-whitsunday-coast-wcs}}
  \item{\url{https://www.qld.gov.au/environment/land/soil/soil-data/survey-types/}}}
}
\usage{
heronvale_known_sites
}
\description{
A dataset containing the locations of described soil profiles within the
Heronvale 1:10000 topographic key map sheet. Sites are from a soil and land
resource survey conducted in the early 2000s. Full profile descriptions can
be viewed in the \href{https://qldglobe.information.qld.gov.au/}{Queensland
Globe}.
}
\keyword{datasets}
