% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{heronvale_soilnames}
\alias{heronvale_soilnames}
\title{Decoded soil class names}
\format{A tibble with 38 rows and 2 columns:
\describe{
\item{CLASS}{Two-letter soil class label}
\item{NAME}{Soil class name}}}
\source{
\itemize{
\item{\url{https://publications.qld.gov.au/dataset/soils-whitsunday-coast-wcs}}
\item{\url{https://www.qld.gov.au/environment/land/soil/soil-data/survey-types/}}}
}
\usage{
heronvale_soilnames
}
\description{
A table of soil class labels found in 'heronvale_soilmap' and 'heronvale_known_sites', along with
their full names. Explanations of the concept associated with each soil name are available in the
 report associated with this mapping project.
}
\keyword{datasets}
