% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsmartr_helpers.R
\name{in_range}
\alias{in_range}
\title{in range}
\usage{
in_range(value = NULL, lower = NULL, upper = NULL, strict = TRUE)
}
\arguments{
\item{value}{Numeric; number to check}

\item{lower}{Numeric; lower limit of range}

\item{upper}{Numeric; upper limit of range}

\item{strict}{Logical; < or <= ?}
}
\value{
TRUE if value is within range
}
\description{
Returns true if number is within range
}
\examples{
in_range(5, 3, 6)
in_range(5, 3, 5, FALSE)
in_range(5, 3, 5, TRUE)
}
