% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsmartr_evaluate.R
\name{n_predicted}
\alias{n_predicted}
\title{Number of classes predicted}
\usage{
n_predicted(input = NULL, n_iterations = NULL, noise_cutoff = NULL)
}
\arguments{
\item{input}{A vector of integers n soil classes long, where the integers refer to the number of
times a soil class was predicted.}

\item{n_iterations}{Integer; the total number of model runs.}

\item{noise_cutoff}{Decimal; proportion of predictions to be considered 'noise' and ignored.
Allowable values range between 0 and 1.}
}
\value{
Integer; the number of distinct soil classes predicted on the pixel (optionally, above
 a supplied threshold).
}
\description{
Counts how many different soil classes were predicted on a pixel
}
\note{
This is a helper function and not widely applicable. It is expected to be used in concert
with \code{\link[raster:calc]{calc}}, applied to a stack of tallied soil class predictions
produced by \code{\link[dsmartr:eval_npred]{eval_npred}}.
}
\examples{
\dontrun{
# a set of tallies from the heronvale demo data
counts_19291 <- as.integer(c(0, 1, 3, 0,  0, 0, 17,  0, 8, 0, 0, 6, 0, 0, 12, 1, 0, 0, 0, 0, 12,
0, 0, 8, 0,  0, 0,  0, 12, 1, 3, 5, 2, 9, 0,  0, 0, 0))
n_soils <- n_soils_predicted(counts_19291)
n_soils_10 <- n_soils_predicted(counts_19291, 100, 0.1)
}

}
\keyword{internal}
