% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsmartr_helpers.R
\name{dsmartr_check_polygons}
\alias{dsmartr_check_polygons}
\title{Check dsmartr polygon inputs}
\usage{
dsmartr_check_polygons(src_map = NULL, id_field = NULL)
}
\arguments{
\item{src_map}{An sfc_POLYGON/MULTIPOLYGON or SpatialPolygonsDataFrame object.}

\item{id_field}{String; name of unique identifier field.}
}
\value{
An sf object with appended logical columns indicating presence of possible data faults.
}
\description{
Checks dsmartr input polygons for potential data quality issues.
}
\details{
This function highlights some issues that may cause dsmartr to either fail or behave
unpredictably. Usually these are side-effects of prepration work carried out in other software.
Sometimes they are the result of errors in the source data. Note that polygon geometry is not
explicitly checked, only attributes. At present, three problems are checked - whether data is
missing (e.g. a polygon class without a matching percentage), whether a percentage is 0%, and
whether total percentage is not 100%.
}
\examples{
\dontrun{
load("heronvale_soilmap")
checked_map <- dsmartr_check_polygons(src_map = heronvale_soilmap,
id_field = 'POLY_NO')}
}
