% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsmartr_unstack.R
\name{dsmartr_class_maps}
\alias{dsmartr_class_maps}
\title{Produce soil class probability surfaces}
\usage{
dsmartr_class_maps(tallied_probs = NULL, soil_class = NULL)
}
\arguments{
\item{tallied_probs}{RasterBrick; \code{tallied_probabilities} output by
\code{\link{dsmartr_collate}}.}

\item{soil_class}{String; Soil class(es) of interest, if only particular maps are desired.}
}
\value{
\code{class_maps}: List of RasterLayers; probability surfaces for each soil class. All
  outputs are written to disk as GeoTIFF.
}
\description{
Generates a soil class probability map for any or all input soil classes. Requires output of
\code{\link{dsmartr_collate}}.
}
\examples{
\dontrun{
# run dsmartr_collate() with the example data and then

# all classes:
classmaps <- dsmartr_class_maps(tallied_probs = collated[['tallied_probabilities']])

# just map two classes of interest:
classmaps <- dsmartr_class_maps(tallied_probs = collated[['tallied_probabilities']],
soil_class = c('An', 'Wr'))}
}
