% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsmartr_evaluate.R
\name{dsmartr_eval_npred}
\alias{dsmartr_eval_npred}
\title{Calculate soil classes per pixel}
\usage{
dsmartr_eval_npred(tallied_preds = NULL, cpus = 1, n_iterations = NULL)
}
\arguments{
\item{tallied_preds}{RasterBrick; \code{tallied_predictions} output by
\code{\link{dsmartr_collate}}.}

\item{cpus}{Integer; number of processors to use in parallel.}

\item{n_iterations}{Integer; the number of iterations that weresupplied to
\code{\link{dsmartr_iterate}}.}
}
\value{
\code{n_classes_predicted}: RasterLayer depicting the number of distinct soils predicted
per pixel. Written to disk as GeoTIFF before being assigned to the global environment.
}
\description{
Calculates and maps the number of distinct soil classes predicted by dsmartr across \code{n}
iterations. Requires outputs of \code{\link{dsmartr_collate}} (with option
\code{keep_tallies = TRUE}).
}
\note{
Fewer classes predicted on a pixel generally indicates higher internal model confidence at
that location.
}
\examples{
\dontrun{
# run dsmartr_collate() with the example data then:
dsmartr_eval_npred(tallied_preds = tallied_predictions,
cpus = max(1, (parallel::detectCores() - 1)),
n_iterations = nlayers(iteration_maps))}
}
