% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsmartr_iterate.R
\name{dsmartr_get_classes}
\alias{dsmartr_get_classes}
\title{Get all soil classes}
\usage{
dsmartr_get_classes(soilmap = NULL, soilpoints = NULL, stub = NULL)
}
\arguments{
\item{soilmap}{Data Frame; Returned by \code{\link{dsmartr_prep_polygons}}.}

\item{soilpoints}{Data Frame; Returned by \code{\link{dsmartr_prep_points}}; optional.}

\item{stub}{String; Common column name or name stub holding soil class data.}
}
\value{
A factor holding the set of soil classes occurring in both datasets, sorted
alphabetically. Internal to \code{\link{dsmartr_iterate}}.
}
\description{
Returns the set of unique soil classes that occur in an input soil map, and, optionally,
a set of known points.
}
\note{
It is up to the user to make sure the classification schema in both datasets matches/is
compatible.
}
\examples{
\dontrun{
# run dsmartr_prep_polygons() and dsmartr_prep_points() with the example code, then:
class_levels <- dsmartr_get_all_classes(soilmap = pr_ap, soilpoints = pr_points, stub = 'CLASS')}
}
\keyword{internal}
