% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsmartr_iterate.R
\name{iter_sample_poly}
\alias{iter_sample_poly}
\title{sample a polygon for dsmartr_iterate}
\usage{
iter_sample_poly(pd = NULL, cs = NULL, ps = NULL, nscol = NULL,
  cellcol = NULL, t_factor = NULL)
}
\arguments{
\item{pd}{A single-row sf object containing soil attributes and geometry. Usually output of
running split(x, 1:nrow(x)) on a polygon sf dataframe.}

\item{cs}{String; stub of attribute column names holding soil class data}

\item{ps}{String; stub of attribute column names folding soil percentage data}

\item{nscol}{String; name of attribute holding number of samples needed}

\item{cellcol}{String; name of attriute holding list of raster cells to sample from}

\item{t_factor}{Integer; dirichlet dampener, supplied by parent function}
}
\value{
A data frame containing two columns: weighted random allocation of soil classes,
and cell numbers
}
\description{
Randomly selects n cells for sampling and assigns them a soil class based on the overlying
map polygon's components
}
\examples{
\dontrun{
# run dsmartr_prep_polygons() and dsmartr_prep_points() with the example code, then:
sample_points <- iter_sample_poly(pd = pr_ap[1, ], cs = 'CLASS', ps = 'PERC',
nscol = 'n_samples', cellcol = 'intersecting_cells', t_factor = t_factor)}
}
\keyword{internal}
