% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsmartr_helpers.R
\name{n_things}
\alias{n_things}
\title{helper function}
\usage{
n_things(input = NULL, selector = NULL)
}
\arguments{
\item{input}{An object of class sfc_POLYGON or sfc_MULTIPOLYGON, see
\code{\link{dsmartr_prepare}}.}

\item{selector}{String; a column name or name stub.}
}
\value{
An atomic vector of non-NA values in the selected columns, by order of occurrence.
}
\description{
Returns cell values from selected columns in an attribute table as an atomic vector,
with NA values excluded.
}
\note{
This is a helper function and not widely applicable. It is expected to be
used by-row on wide-formatted sf attribute data.
}
\examples{
library(dplyr)
library(tibble)
ex <- tribble( ~ID, ~CLASS_1, ~CLASS_2, ~PERC_1, ~PERC_2,
                 1,      'A',      'B',      60,      40)
n_things(ex, 'PERC')
}
\keyword{internal}
